Convert karma configs to typescript
We can do that even in older node versions. Better than cjs configs, eh?
