+++ packages/proto-rpc/package.json
@@ -25,7 +25,7 @@
     "clean": "jest --clearCache || true; rm -rf dist *.tsbuildinfo node_modules/.cache || true",
     "eslint": "./test-proto.sh && eslint --cache --cache-location=node_modules/.cache/.eslintcache/ '*/**/*.{js,ts}'",
     "test": "./test-proto.sh && npm run test-unit && npm run test-integration",
-    "test-browser": "./test-proto.sh && karma start karma.config.js",
+    "test-browser": "./test-proto.sh && karma start karma.config.ts",
     "test-unit": "./test-proto.sh && jest test/unit",
     "test-integration": "./test-proto.sh && jest --bail test/integration"
   },
