+++ packages/sdk/karma-integration.config.ts
@@ -1,0 +1,4 @@
+import webpackConfig from './webpack-karma.config'
+import { createKarmaConfig } from '@streamr/browser-test-runner'
+
+export default createKarmaConfig(['test/integration/**/*.ts'], webpackConfig, __dirname)
