--- packages/trackerless-network/karma.config.js
@@ -1,30 +1,0 @@
-/* eslint-disable @typescript-eslint/no-require-imports */
-const path = require('path')
-const { createKarmaConfig, createWebpackConfig } = require('@streamr/browser-test-runner')
-
-const TEST_PATHS = [
-    './test/unit/**/*.ts',
-    './test/integration/**/*.ts',
-    './test/end-to-end/**/!(webrtc*|websocket*)',
-]
-
-module.exports = createKarmaConfig(TEST_PATHS, createWebpackConfig({
-    libraryName: 'trackerless-network',
-    alias: {
-        [path.resolve(__dirname, '../dht/src/connection/webrtc/NodeWebrtcConnection.ts')]:
-            path.resolve(__dirname, '../dht/src/connection/webrtc/BrowserWebrtcConnection.ts'),
-        [path.resolve(__dirname, '../dht/src/connection/websocket/NodeWebsocketClientConnection.ts')]:
-            path.resolve(__dirname, '../dht/src/connection/websocket/BrowserWebsocketClientConnection.ts'),
-        '@streamr/dht': path.resolve('../dht/src/exports.ts'),
-        '@streamr/proto-rpc': path.resolve('../proto-rpc/src/exports.ts'),
-    },
-    fallback: {
-        module: false
-    },
-    externals: {
-        'node-datachannel': 'commonjs node-datachannel',
-        express: 'Express',
-        http: 'HTTP',
-        ws: 'WebSocket',
-    }
-}), __dirname)
