--- packages/utils/karma.config.js
@@ -1,11 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-const { createKarmaConfig, createWebpackConfig } = require('@streamr/browser-test-runner')
-
-const TEST_PATHS = ['test/**/*.ts']
-
-module.exports = createKarmaConfig(TEST_PATHS, createWebpackConfig({
-    libraryName: 'utils',
-    fallback: {
-        module: false
-    }
-}))
