+++ packages/utils/package.json
@@ -21,7 +21,7 @@
     "clean": "jest --clearCache || true; rm -rf dist *.tsbuildinfo node_modules/.cache || true",
     "eslint": "eslint --cache --cache-location=node_modules/.cache/.eslintcache/ '*/**/*.{js,ts}'",
     "test": "jest",
-    "test-browser": "karma start karma.config.js"
+    "test-browser": "karma start karma.config.ts"
   },
   "author": "Streamr Network AG <contact@streamr.network>",
   "license": "Apache-2.0",
